requirement.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
click==7.1.1
cryptography==2.3.1
cycler==0.10.0
et-xmlfile==1.0.1
Flask==1.1.2
idna==2.9
imutils==0.5.3
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.0
numpy==1.14.2
olefile==0.46
opencv-python==4.2.0.34
openpyxl==2.6.4
pandas==0.23.4
Pillow==5.2.0
pycparser==2.20
pyOpenSSL==18.0.0
pyparsing==2.4.7
PySocks==1.6.8
python-dateutil==2.8.1
pytz==2019.3
six==1.11.0
tornado==5.1.1
urllib3==1.23
Werkzeug==1.0.1
win-inet-pton==1.0.1
wincertstore==0.2
xlrd==1.2.0
